repeat timer command
