+++ lab3/include/irq.h
@@ -1,5 +1,4 @@
 #ifndef	_IRQ_H
 #define	_IRQ_H
 void handle_irq(void);
-void enable_interrupt_controller();
 #endif
