+++ lab3/include/peripherals/irq.h
@@ -20,4 +20,5 @@
 #define SYSTEM_TIMER_IRQ_3	(1 << 3)
 
 #define CORE0_INTERRUPT_SOURCE 0x40000060
+#define AUX_IRQ (1 << 29)
 #endif  /*_P_IRQ_H */
