+++ lab3/src/sys.S
@@ -13,3 +13,4 @@ sync_call_exc:
 sync_call_uart:
     svc #3
     ret
+
